requirements.txt
fastapi
uvicorn
apscheduler
jinja2
python-dotenv
